web/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.5.1
python-dotenv==1.0.1
pymongo==4.3.3
setuptools==57.4.0
passlib==1.7.4
jwt==1.3.1
authlib==1.3.0
httpx==0.27.0